refactor: v1 v2 for models and services
added v1 and v2 folders for models and services

fixed code for imports for services, models, controllers, routes
