+++ backend/src/services/v1/transactionService.js
@@ -1,4 +1,4 @@
-const Transaction = require("../models/Transaction"); // Ensure this matches your filename
+const Transaction = require("../../models/v1/Transaction"); // Ensure this matches your filename
 
 class TransactionService {
   /**
